% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelpbeds-package.R
\docType{data}
\name{ipadic}
\alias{ipadic}
\title{Vocabularies of 'IPAdic'}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 392126 rows and 13 columns.
}
\usage{
ipadic
}
\description{
Vocabularies of 'IPAdic'
}
\details{
A dataset containing vocabularies of 'IPAdic'.

The original 'IPAdic' is copyrighted
by the Nara Institute of Science and Technology and 'ICOT'.
For details about the terms of use, see 'LICENSE.note' in this package.
}
\keyword{datasets}
