% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelpbeds-package.R
\docType{data}
\name{ipadicmat}
\alias{ipadicmat}
\title{'matrix.def' of 'IPAdic'}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1731856 rows and 3 columns.
}
\usage{
ipadicmat
}
\description{
'matrix.def' of 'IPAdic'
}
\details{
A dataset converted from 'matrix.def' of 'IPAdic'.

The original 'IPAdic' is copyrighted
by the Nara Institute of Science and Technology and 'ICOT'.
For details about the terms of use, see 'LICENSE.note' in this package.
}
\keyword{datasets}
