% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelpbeds-package.R
\docType{package}
\name{kelpbeds-package}
\alias{kelpbeds}
\alias{kelpbeds-package}
\title{kelpbeds: Dictionary Tool for 'MeCab'}
\description{
Provides the source 'IPAdic' for 'MeCab'.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Other contributors:
\itemize{
  \item Nara Institute of Science and Technology [copyright holder]
  \item ICOT Free Software [copyright holder]
}

}
\keyword{internal}
