% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-ipadic.R
\name{prep_ipadic}
\alias{prep_ipadic}
\title{Prepare source IPA dictionary}
\usage{
prep_ipadic(dic_dir = tempdir())
}
\arguments{
\item{dic_dir}{Directory to write dictionary files.}
}
\value{
\code{dic_dir} is returend invisibly.
}
\description{
Prepare source IPA dictionary
}
