% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-skkserv.R
\name{prep_skkserv}
\alias{prep_skkserv}
\title{Prepare source MeCab SKKServ dictionary}
\usage{
prep_skkserv(dic_dir = tempdir())
}
\arguments{
\item{dic_dir}{Directory to write dictionary files.}
}
\value{
\code{dic_dir} is returend invisibly.
}
\description{
Prepare source MeCab SKKServ dictionary
}
