% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-sudachidict.R
\name{prep_sudachidict}
\alias{prep_sudachidict}
\title{Prepare directory for building 'Sudachi' dictionary}
\usage{
prep_sudachidict(dir = tempdir())
}
\arguments{
\item{dir}{Directory to write dictionary files.}
}
\value{
\code{file.path(dir, "sudachidict")} is returend invisibly.
}
\description{
Prepare directory for building 'Sudachi' dictionary
}
