% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelpbeds-package.R
\docType{data}
\name{skkservmat}
\alias{skkservmat}
\title{'matrix.def' of 'mecab-sskserv' dictionary}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 897981 rows and 3 columns.
}
\usage{
skkservmat
}
\description{
'matrix.def' of 'mecab-sskserv' dictionary
}
\details{
A dataset converted from 'matrix.def' of 'mecab-sskserv' dictionary.

The dictionary of 'mecab-sskserv' is a derived work from 'IPAdic'
and licensed under the same license as 'IPAdic'.

The original 'IPAdic' is copyrighted
by the Nara Institute of Science and Technology and 'ICOT'.
For details about the terms of use, see 'LICENSE.note' in this package.
}
\keyword{datasets}
